win32 fixes after gl renderer refactor on osx, win32 volume enumeration, fs enumeration started

+++ examples/code/input/input_example.cpp
@@ -78,7 +78,7 @@ PEN_THREAD_RETURN pen::game_entry( void* params )
         const pen::mouse_state& ms = pen::input_get_mouse_state( );
         
         //mouse
-        vec2f mouse_pos = vec2f( ms.x, vp.height - ms.y );
+        vec2f mouse_pos = vec2f( (f32)ms.x, vp.height - (f32)ms.y );
         vec2f mouse_quad_size = vec2f( 5.0f, 5.0f );
         dbg::add_quad_2f( mouse_pos, mouse_quad_size, vec3f::cyan() );
